requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
pydantic==2.11.3
pydantic-settings==2.9.1
python-dotenv==1.1.0
bcrypt==4.0.1
requests==2.32.3
mysql-connector-python==8.4.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
twilio==9.0.4


